TVL and APR updated to support changed made by individual claim feature
